# healthmanagementapi/Dtos/HealthFacilityDto.cs
namespace healthmanagementapi.Dtos
{
    public class HealthFacilityDto
    {
        public int? Id { get; set; }
        public string Name { get; set; }
        public string District { get; set; }
        public string Region { get; set; }
        public string State { get; set; }
        public string Country { get; set; }
    }

}