# healthmanagementapi/Dtos/LoginModel.cs
using healthmanagementapi.Entities;

namespace healthmanagementapi.Dtos
{
    public class LoginModel
    {
        public string Email { get; set; }
        public string Password { get; set; }
    }

}