# healthmanagementapi/Dtos/RegisterModel.cs
namespace healthmanagementapi.Dtos
{
    public class RegisterModel
    {
        public string Email { get; set; }
        public string Password { get; set; }
        public string Name { get; set; }
    }

}